+++ packages/cuttlefish/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @typecad/cuttlefish
 
+## 1.0.0-alpha.8
+
+### Patch Changes
+
+- @typecad/ui@1.0.0-alpha.8
+- @typecad/safety@1.0.0-alpha.8
+
 ## 1.0.0-alpha.7
 
 ### Minor Changes