+++ packages/expect/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @typecad/expect
 
+## 1.0.0-alpha.8
+
+### Patch Changes
+
+- @typecad/arduino-cli@1.0.0-alpha.8
+
 ## 1.0.0-alpha.7
 
 ### Patch Changes