+++ packages/framework-arduino/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @typecad/framework-arduino
 
+## 1.0.0-alpha.8
+
+### Patch Changes
+
+- @typecad/cuttlefish@1.0.0-alpha.8
+- @typecad/hal@1.0.0-alpha.8
+- @typecad/arduino-cli@1.0.0-alpha.8
+
 ## 1.0.0-alpha.7
 
 ### Patch Changes