+++ packages/framework-native/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @typecad/framework-native
 
+## 1.0.0-alpha.8
+
+### Patch Changes
+
+- @typecad/cuttlefish@1.0.0-alpha.8
+
 ## 1.0.0-alpha.7
 
 ### Patch Changes