+++ packages/simulator/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @typecad/simulator
 
+## 1.0.0-alpha.8
+
+### Patch Changes
+
+- @typecad/hal@1.0.0-alpha.8
+
 ## 1.0.0-alpha.7
 
 ### Patch Changes